ArrayIndexing: Split utility collection into a separate sub-package
