+++ loki/transformations/array_indexing/__init__.py
@@ -1,0 +1,15 @@
+# (C) Copyright 2018- ECMWF.
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+"""
+Collection of utility routines to deal with common array indexing conversions.
+"""
+
+from loki.transformations.array_indexing.array_indices import * # noqa
+from loki.transformations.array_indexing.demote import * # noqa
+from loki.transformations.array_indexing.promote import * # noqa
+from loki.transformations.array_indexing.vector_notation import * # noqa
